requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.10.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
biosppy==1.0.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
cmaes==0.9.1
colorlog==6.7.0
comm==0.1.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
einops==0.7.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
greenlet==2.0.2
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
iterative-stratification==0.1.7
jedi==0.19.0
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-client==8.3.1
jupyter-console==6.6.3
jupyter-core==5.3.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-server==2.7.2
jupyter-server-terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.24.0
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
lightning-utilities==0.9.0
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==3.0.1
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.2
notebook-shim==0.2.3
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.8.0.76
opt-einsum==3.3.0
optuna==3.1.0
overrides==7.4.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.0.9.post0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.4.0
referencing==0.30.2
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.0
scikit-learn==1.2.1
scipy==1.10.1
Send2Trash==1.8.2
shortuuid==1.0.11
six==1.16.0
slack-sdk==3.20.1
sniffio==1.3.0
soundfile==0.11.0
soupsieve==2.4.1
SQLAlchemy==2.0.5.post1
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
# torch==1.12.1
# torchaudio==0.12.1
torchdiffeq==0.2.3
torchmetrics==1.2.0
# torchvision==0.13.1
tornado==6.3.3
tqdm==4.64.1
traitlets==5.9.0
typing-extensions==4.5.0
uri-template==1.3.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
wfdb==3.3.0
widgetsnbextension==4.0.8
yarl==1.9.2
zipp==3.15.0
